prac: valid parentheses mapping solution reasoning
